handle password login error
