+++ src/app/pages/auth/Login.tsx
@@ -1,276 +1,19 @@
-import React, { FormEventHandler, useCallback, useState } from 'react';
-import {
-  Box,
-  Button,
-  Header,
-  Icon,
-  IconButton,
-  Icons,
-  Input,
-  Menu,
-  PopOut,
-  Text,
-  color,
-  config,
-} from 'folds';
-import FocusTrap from 'focus-trap-react';
+import React from 'react';
+import { Box, Text, color } from 'folds';
 import { Link, generatePath, useSearchParams } from 'react-router-dom';
-import { LoginRequest, createClient } from 'matrix-js-sdk';
-import to from 'await-to-js';
 import { REGISTER_PATH } from '../paths';
 import { useAuthFlows } from '../../hooks/useAuthFlows';
 import { useAuthServer } from '../../hooks/useAuthServer';
-import { UseStateProvider } from '../../components/UseStateProvider';
 import { useParsedLoginFlows } from '../../hooks/useParsedLoginFlows';
-import { getMxIdLocalPart, getMxIdServer, isUserId } from '../../utils/matrix';
-import { EMAIL_REGEX } from '../../utils/regex';
-import { useAutoDiscoveryInfo } from '../../hooks/useAutoDiscoveryInfo';
-import { useAsyncCallback } from '../../hooks/useAsyncCallback';
-import { autoDiscovery, specVersions } from '../../cs-api';
+import { PasswordLoginForm } from './PasswordLoginForm';
 
-function UsernameHint({ server }: { server: string }) {
-  const [open, setOpen] = useState(false);
-  return (
-    <PopOut
-      open={open}
-      position="Top"
-      align="End"
-      content={
-        <FocusTrap
-          focusTrapOptions={{
-            initialFocus: false,
-            onDeactivate: () => setOpen(false),
-            clickOutsideDeactivates: true,
-          }}
-        >
-          <Menu>
-            <Header size="300" style={{ padding: `0 ${config.space.S200}` }}>
-              <Text size="L400">Hint</Text>
-            </Header>
-            <Box
-              style={{ padding: config.space.S200, paddingTop: 0 }}
-              direction="Column"
-              tabIndex={0}
-              gap="100"
-            >
-              <Text size="T300">
-                <Text as="span" size="Inherit" priority="300">
-                  Username:
-                </Text>{' '}
-                johndoe
-              </Text>
-              <Text size="T300">
-                <Text as="span" size="Inherit" priority="300">
-                  Matrix ID:
-                </Text>
-                {` @johndoe:${server}`}
-              </Text>
-              <Text size="T300">
-                <Text as="span" size="Inherit" priority="300">
-                  Email:
-                </Text>
-                {` johndoe@${server}`}
-              </Text>
-            </Box>
-          </Menu>
-        </FocusTrap>
-      }
-    >
-      {(targetRef) => (
-        <IconButton
-          tabIndex={-1}
-          onClick={() => setOpen(true)}
-          ref={targetRef}
-          type="button"
-          variant="Background"
-          size="400"
-          radii="300"
-        >
-          <Icon style={{ opacity: config.opacity.P300 }} size="100" src={Icons.Info} />
-        </IconButton>
-      )}
-    </PopOut>
-  );
-}
-
-type PasswordLoginFormProps = {
-  defaultUsername?: string;
-  defaultEmail?: string;
-};
-export function PasswordLoginForm({ defaultUsername, defaultEmail }: PasswordLoginFormProps) {
-  const server = useAuthServer();
-
-  const serverDiscovery = useAutoDiscoveryInfo();
-  const baseUrl = serverDiscovery['m.homeserver'].base_url;
-
-  const [loginState, startLogin] = useAsyncCallback(
-    useCallback(async (serverBaseUrl: string, data: Omit<LoginRequest, 'type'>) => {
-      const mx = createClient({ baseUrl: serverBaseUrl });
-      // const [err, res] = await to(mx.login('m.login.password', data));
-      // return res;
-    }, [])
-  );
-
-  const handleUsernameLogin = (username: string, password: string) => {
-    startLogin(baseUrl, {
-      identifier: {
-        type: 'm.id.user',
-        user: username,
-      },
-      password,
-      initial_device_display_name: 'Cinny Web',
-    });
-  };
-  const handleMxIdLogin = async (mxId: string, password: string) => {
-    const mxIdServer = getMxIdServer(mxId);
-    const mxIdUsername = getMxIdLocalPart(mxId);
-    if (!mxIdServer || !mxIdUsername) return;
-
-    const [, discovery] = await to(autoDiscovery(fetch, mxIdServer));
-
-    let mxIdBaseUrl: string | undefined;
-    const [, discoveryInfo] = discovery ?? [];
-
-    if (discoveryInfo) {
-      mxIdBaseUrl = discoveryInfo['m.homeserver'].base_url;
-    }
-    // TODO: add displayAlertDialog func
-    // which will create another react root with portal
-    if (!mxIdBaseUrl) {
-      alert(
-        'Failed to find MXID homeserver! Please enter server in Homeserver input for more details.'
-      );
-      return;
-    }
-    const [, versions] = await to(specVersions(fetch, mxIdBaseUrl));
-    if (!versions) {
-      alert('Homeserver URL does not appear to be a valid Matrix homeserver.');
-      return;
-    }
-
-    startLogin(mxIdBaseUrl, {
-      identifier: {
-        type: 'm.id.user',
-        user: mxIdUsername,
-      },
-      password,
-      initial_device_display_name: 'Cinny Web',
-    });
-  };
-  const handleEmailLogin = (email: string, password: string) => {
-    startLogin(baseUrl, {
-      identifier: {
-        type: 'm.id.thirdparty',
-        medium: 'email',
-        address: email,
-      },
-      password,
-      initial_device_display_name: 'Cinny Web',
-    });
-  };
-
-  const handleSubmit: FormEventHandler<HTMLFormElement> = (evt) => {
-    evt.preventDefault();
-    const { usernameInput, passwordInput } = evt.target as HTMLFormElement & {
-      usernameInput: HTMLInputElement;
-      passwordInput: HTMLInputElement;
-    };
-
-    const username = usernameInput.value.trim();
-    const password = passwordInput.value;
-    if (!username) {
-      usernameInput.focus();
-      return;
-    }
-    if (!password) {
-      passwordInput.focus();
-      return;
-    }
-
-    if (isUserId(username)) {
-      handleMxIdLogin(username, password);
-      return;
-    }
-    if (EMAIL_REGEX.test(username)) {
-      handleEmailLogin(username, password);
-      return;
-    }
-    handleUsernameLogin(username, password);
-  };
-
-  return (
-    <Box as="form" onSubmit={handleSubmit} direction="Inherit" gap="400">
-      <Box direction="Column" gap="100">
-        <Text as="label" size="L400" priority="300">
-          Username
-        </Text>
-        <Input
-          defaultValue={defaultUsername ?? defaultEmail}
-          style={{ paddingRight: config.space.S200 }}
-          name="usernameInput"
-          variant="Background"
-          size="500"
-          required
-          outlined
-          after={<UsernameHint server={server} />}
-        />
-      </Box>
-      <Box direction="Column" gap="100">
-        <Text as="label" size="L400" priority="300">
-          Password
-        </Text>
-        <UseStateProvider initial={false}>
-          {(visible, setVisible) => (
-            <Input
-              style={{ paddingRight: config.space.S200 }}
-              name="passwordInput"
-              type={visible ? 'text' : 'password'}
-              variant={visible ? 'Warning' : 'Background'}
-              size="500"
-              outlined
-              required
-              after={
-                <IconButton
-                  onClick={() => setVisible(!visible)}
-                  type="button"
-                  variant={visible ? 'Warning' : 'Background'}
-                  size="400"
-                  radii="300"
-                >
-                  <Icon
-                    style={{ opacity: config.opacity.P300 }}
-                    size="100"
-                    src={visible ? Icons.Eye : Icons.EyeBlind}
-                  />
-                </IconButton>
-              }
-            />
-          )}
-        </UseStateProvider>
-        <Box justifyContent="End">
-          <Text as="span" size="T200" priority="300" align="Right">
-            {/* TODO: make reset password path */}
-            <Link to={generatePath(REGISTER_PATH, { server })}>Forget Password?</Link>
-          </Text>
-        </Box>
-      </Box>
-      <span />
-      <Button type="submit" variant="Primary" size="500">
-        <Text as="span" size="B500">
-          Login
-        </Text>
-      </Button>
-    </Box>
-  );
-}
-
-const getLoginSearchParams = (
-  searchParams: URLSearchParams
-): {
+export type LoginSearchParams = {
   username?: string;
   email?: string;
   loginToken?: string;
-} => ({
+};
+
+const getLoginSearchParams = (searchParams: URLSearchParams): LoginSearchParams => ({
   username: searchParams.get('username') ?? undefined,
   email: searchParams.get('email') ?? undefined,
   loginToken: searchParams.get('loginToken') ?? undefined,
